requirements.txt
# API
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.20

# Settings
pydantic-settings==2.7.0

# Database
sqlalchemy==2.0.36
asyncpg==0.30.0
pgvector==0.3.6

# Face Recognition
insightface==0.7.3
onnxruntime==1.20.1        # CPU; use onnxruntime-gpu para GPU
numpy==1.26.4
Pillow==11.0.0
opencv-python-headless==4.10.0.84

# Liveness (AWS)
boto3==1.35.81

# Logging
structlog==24.4.0

# Dev / Test
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1
